OpenGL client enhanced, looks like the web app can not handle the performance

+++ lolirift/Controllers/MapController.cs
@@ -40,7 +40,7 @@ namespace lolirift.Controllers
                 controller = "map",
                 width = grid.Width,
                 height = grid.Height,
-                heightmap = grid.Grid.Select(f => f.Height).ToArray()
+                heightmap = grid.Grid.Select(f => (int)f.Height).ToArray()
             };
 
             var json = JsonConvert.SerializeObject(response);
